backend/pyproject.toml
[project]
name = "mealplanner-backend"
version = "0.1.0"
description = "Mealplanner 2.0 backend"
requires-python = ">=3.11"
dependencies = [
    "dlt[duckdb]>=1.4.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "dbt-duckdb>=1.8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-ai-slim[openai]>=1.73.0",
]

[project.optional-dependencies]
postgres = ["dlt[postgres]>=1.4.0"]
dev = ["ruff>=0.4.0", "pytest>=8.0.0", "httpx>=0.27.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["pipeline", "api"]
